Switched for integer to fractional coords for pixels
