+++ python/pyenki/viewer/utils.py
@@ -1,17 +1,21 @@
+from __future__ import annotations
+
 import typing
 
 import numpy as np
 import numpy.typing
 from OpenGL import GL  # type: ignore[import-untyped]
 from PySide6.QtCore import QCoreApplication, QEventLoop, Qt, QTimer
-from PySide6.QtGui import (QImage, QOpenGLContext, QOpenGLFunctions,
-                           QSurfaceFormat, QVector4D)
+from PySide6.QtGui import QImage, QOpenGLContext, QSurfaceFormat, QVector4D
 from PySide6.QtWidgets import QApplication
 
-from .. import Image, PhysicalObject, VectorLike, World
+from .. import Image
 from .camera import Camera
 from .renderer import Renderer
 from .types import Pixel, Vector3
+
+if typing.TYPE_CHECKING:
+    from PySide6.QtGui import QOpenGLFunctions
 
 
 def init(share: bool = True) -> None:
@@ -69,16 +73,6 @@ def functions() -> QOpenGLFunctions:
     return QOpenGLContext.currentContext().functions()
 
 
-def get_object_at(world: World,
-                  position: VectorLike,
-                  tolerance: float = 0) -> PhysicalObject | None:
-    position = np.asarray(position)
-    for obj in world.objects:
-        if obj.contains(position, tolerance):
-            return obj
-    return None
-
-
 def get_position_of_pixel(pixel: Pixel, width: int, height: int,
                           camera: Camera) -> Vector3 | None:
     depth = get_depth_of_pixel(pixel, width, height)
@@ -86,8 +80,8 @@ def get_position_of_pixel(pixel: Pixel, width: int, height: int,
         return None
     m = camera.projection * camera.matrix
     m, _ = m.inverted()
-    x = (pixel[0] - width * 0.5) / (width * 0.5)
-    y = (height - pixel[1] - height * 0.5) / (height * 0.5)
+    x = 2 * pixel[0] - 1
+    y = 1 - 2 * pixel[1]
     p = m.map(QVector4D(x, y, 2 * depth - 1, 1))
     if p.w():
         return np.array((p.x(), p.y(), p.z())) / p.w()
@@ -95,17 +89,19 @@ def get_position_of_pixel(pixel: Pixel, width: int, height: int,
 
 
 def get_depth_of_pixel(pixel: Pixel, width: int, height: int) -> float | None:
+    if pixel[0] < 0 or pixel[1] < 0 or pixel[0] > 1 or pixel[1] > 1:
+        return None
     f = functions()
     data = np.zeros(1, np.float32)
-    if pixel[0] < width and pixel[1] < height:
-        f.glReadPixels(
-            pixel[0],
-            height - pixel[1],
-            1,
-            1,
-            GL.GL_DEPTH_COMPONENT,
-            GL.GL_FLOAT,
-            data.data  # type: ignore
-        )
-        return float(data[0])
-    return None
+    i = round(pixel[0] * (width - 1))
+    j = round((1 - pixel[1]) * (height - 1))
+    f.glReadPixels(
+        i,
+        j,
+        1,
+        1,
+        GL.GL_DEPTH_COMPONENT,
+        GL.GL_FLOAT,
+        data.data  # type: ignore
+    )
+    return float(data[0])
